exported hbs helpers to a file and cleaned css
