+++ src/handlebars/helpers.ts
@@ -1,0 +1,10 @@
+import { handlebars } from 'hbs';
+
+export function registerHandlebarsHelpers() {
+    handlebars.registerHelper('ifCond', function (v1, v2, options) {
+        if (v1 === v2) {
+            return options.fn(this);
+        }
+        return options.inverse(this);
+    });
+}
